fixed : compile warnings when using stlport std::min and max they required implicit cast on result
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@3233 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/Object.cpp
@@ -608,7 +608,7 @@ void Object::_BuildValuesUpdate(ByteBuffer * data, UpdateMask *updateMask, Playe
 	if( m_valuesCount > ( 2 * 0x20 ) )//if number of blocks > 2->  unit and player+item container
 	{
 		bc = updateMask->GetUpdateBlockCount();
-		values_count = min( bc * 32, m_valuesCount );
+		values_count = (uint32)min( bc * 32, m_valuesCount );
 
 	}
 	else
